Make :include public on I18n::Backend::Simple
Rationale: the I18n::Backend part of this library is designed to provide various modules that can be included into the I18n::Backend::Simple backend or other compatible backend implementations and thereby extend their functionality in an optional manner.

Thus, both code and documentation have been cluttered with lots of calls sending :include to the Simple backend class. It seems that we in fact treat sending include as a public API of the Simple backend class and therefor should make it public.